pyproject.toml
[project]
name = "ai-mcp-net-analysis"
version = "0.1.0"
description = "nmap based Model Context Protocol module"
readme = "README.md"
requires-python = ">=3.13"

authors = [
    { name = "zencodeai", email = "info@payzensecurity.ca" },
]

keywords = ["nmap", "network", "mcp", "llm"]
license = { text = "MIT" }

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "pyyaml>=6.0.2",
    "schema>=0.7.7",
]

[project.scripts]
ai-mcp-net-analysis = "ai_mcp_net_analysis:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "freezegun>=1.5.1",
    "pyright>=1.1.389",
    "pytest>=8.3.3",
    "ruff>=0.8.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_mcp_net_analysis"]
